requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
python-multipart
psycopg2-binary
fastapi-jwt
bcrypt
pydantic
python-dotenv
python-jose
SQLAlchemy